+++ src/plugins/janadot/JEventProcessorJANADOT.h
@@ -75,8 +75,8 @@ class JEventProcessorJANADOT:public JEventProcessor
 		const char* className(void){return "JEventProcessorJANADOT";}
 
 		jerror_t init(void);							///< Called once at program start.
-		jerror_t brun(JEventLoop *loop, int runnumber);	///< Called everytime a new run number is detected.
-		jerror_t evnt(JEventLoop *loop, int eventnumber);	///< Called every event.
+		jerror_t brun(JEventLoop *loop, int32_t runnumber);	///< Called everytime a new run number is detected.
+		jerror_t evnt(JEventLoop *loop, uint64_t eventnumber);	///< Called every event.
 		jerror_t erun(void){return NOERROR;};	///< Called everytime run number changes, provided brun has been called.
 		jerror_t fini(void);							///< Called after last event of last event source has been processed.
